==> 5. Unione finale stazioni.ipynb <==
import pandas as pd
import numpy as np

stazioni_vt = pd.read_csv('stazioni_vt.csv')

stazioni_rfi_binari = pd.read_json('stazioni_binari.json', orient = 'index')
stazioni_rfi_binari.reset_index(inplace=True)
stazioni_rfi_binari.rename(columns={'index':'nome_stazione_rfi', 0:'binari'}, inplace=True)
stazioni_rfi_binari['binari'] = stazioni_rfi_binari['binari'].str.replace(' ', '', regex=False)

merging_table = pd.read_csv('matching stazioni_vt-stazioni_rfi.csv', index_col=0)

stazioni_wiki = pd.read_csv('binari wiki.csv', index_col=0)
stazioni_wiki = stazioni_wiki['binari_ok']

stazioni_finale = pd.merge(stazioni_vt, merging_table, left_index=True, right_on='id_vt', how='outer')
stazioni_finale = pd.merge(stazioni_finale, stazioni_rfi_binari, left_on='id_rfi', right_index=True, how='left')
stazioni_finale.drop(columns=['id_vt', 'id_rfi', 'somigl_nomi_stazioni','nome_stazione_rfi'], inplace=True)
stazioni_finale = pd.merge(stazioni_finale, stazioni_wiki, left_on='id', right_index=True, how='left')

stazioni_finale['binari_finale'] = stazioni_finale['binari'].fillna('').astype(str)+stazioni_finale['binari_ok'].fillna('').astype(str)
stazioni_finale['binari_finale'] = stazioni_finale['binari_finale'].str.replace(r'^\s*$', 'NaN', regex=True)
stazioni_finale['binari_finale'] = stazioni_finale['binari_finale'].astype(dtype='float64')
stazioni_finale['binari_finale'] = stazioni_finale['binari_finale'].astype(dtype='Int64')
stazioni_finale.drop(columns=['binari', 'binari_ok'], inplace=True)
stazioni_finale.set_index('id', drop=True, inplace=True)
stazioni_finale.to_csv('stazioni_finale.csv')
